Moved all of the UI components into their own directory.
